# public/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ORCA DAM API",
    "description": "Digital Asset Management API for ORCA DAM. Provides endpoints for managing assets, tags, and chunked file uploads.\n\n## Authentication\n\nAll endpoints except `/api/assets/meta` require authentication via Bearer token (Laravel Sanctum or JWT).\n\nInclude the token in the Authorization header:\n```\nAuthorization: Bearer YOUR_TOKEN_HERE\n```\n\n## Rate Limiting\n\nChunked upload endpoints are rate-limited to 100 requests per minute.",
    "version": "1.0.0",
    "contact": {
      "name": "ORCA DAM"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/assets/search": {
      "get": {
        "tags": ["Assets"],
        "summary": "Search assets",
        "description": "Search assets by query, tags, and type. Optimized for asset picker integrations.",
        "operationId": "searchAssets",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query (filename)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Filter by tag IDs (comma-separated)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by file type",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["images", "videos", "documents"]
            }
          },
          {
            "name": "folder",
            "in": "query",
            "description": "Filter by S3 folder path (e.g., 'assets/marketing')",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of items per page (max 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 24,
              "maximum": 100
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order for results",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["date_desc", "date_asc", "upload_desc", "upload_asc", "size_desc", "size_asc", "name_asc", "name_desc", "s3key_asc", "s3key_desc"],
              "default": "date_desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAssets"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/tags": {
      "get": {
        "tags": ["Tags"],
        "summary": "List tags",
        "description": "Get all tags with optional type filtering",
        "operationId": "listTags",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter by tag type",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["user", "ai"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/tags/search": {
      "get": {
        "tags": ["Tags"],
        "summary": "Search tags",
        "description": "Search tags by name",
        "operationId": "searchTags",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by tag type",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["user", "ai"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Laravel Sanctum API token"
      }
    },
    "schemas": {
      "Asset": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "s3_key": {
            "type": "string",
            "example": "assets/abc123-uuid.jpg"
          },
          "filename": {
            "type": "string",
            "example": "beach-sunset.jpg"
          },
          "mime_type": {
            "type": "string",
            "example": "image/jpeg"
          },
          "size": {
            "type": "integer",
            "example": 1048576,
            "description": "File size in bytes"
          },
          "width": {
            "type": "integer",
            "nullable": true,
            "example": 1920
          },
          "height": {
            "type": "integer",
            "nullable": true,
            "example": 1080
          },
          "thumbnail_s3_key": {
            "type": "string",
            "nullable": true,
            "example": "thumbnails/abc123-uuid_thumb.jpg"
          },
          "resize_s_s3_key": {
            "type": "string",
            "nullable": true,
            "example": "thumbnails/S/abc123-uuid.jpg",
            "description": "S3 key for small resize variant"
          },
          "resize_m_s3_key": {
            "type": "string",
            "nullable": true,
            "example": "thumbnails/M/abc123-uuid.jpg",
            "description": "S3 key for medium resize variant"
          },
          "resize_l_s3_key": {
            "type": "string",
            "nullable": true,
            "example": "thumbnails/L/abc123-uuid.jpg",
            "description": "S3 key for large resize variant"
          },
          "alt_text": {
            "type": "string",
            "nullable": true,
            "example": "Beautiful sunset over the ocean"
          },
          "caption": {
            "type": "string",
            "nullable": true,
            "example": "Sunset at Malibu Beach"
          },
          "license_type": {
            "type": "string",
            "nullable": true,
            "example": "cc_by"
          },
          "license_expiry_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "copyright": {
            "type": "string",
            "nullable": true,
            "example": "John Doe"
          },
          "copyright_source": {
            "type": "string",
            "nullable": true
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Public S3 URL for the asset"
          },
          "thumbnail_url": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "Public S3 URL for the thumbnail"
          },
          "resize_s_url": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "Public URL for small resize variant"
          },
          "resize_m_url": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "Public URL for medium resize variant"
          },
          "resize_l_url": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "Public URL for large resize variant"
          },
          "formatted_size": {
            "type": "string",
            "example": "1.0 MB",
            "description": "Human-readable file size"
          },
          "folder": {
            "type": "string",
            "example": "assets",
            "description": "S3 folder path"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "sunset"
          },
          "type": {
            "type": "string",
            "enum": ["user", "ai"],
            "example": "user"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "User": {
        "type": "object",
        "description": "User object (sensitive fields like email are hidden from API responses)",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "role": {
            "type": "string",
            "enum": ["editor", "admin", "api"],
            "example": "editor"
          }
        }
      },
      "PaginatedAssets": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "integer",
            "example": 1
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          },
          "first_page_url": {
            "type": "string",
            "format": "uri"
          },
          "from": {
            "type": "integer",
            "nullable": true
          },
          "last_page": {
            "type": "integer"
          },
          "last_page_url": {
            "type": "string",
            "format": "uri"
          },
          "links": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "format": "uri",
                  "nullable": true
                },
                "label": {
                  "type": "string"
                },
                "active": {
                  "type": "boolean"
                }
              }
            }
          },
          "next_page_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "path": {
            "type": "string",
            "format": "uri"
          },
          "per_page": {
            "type": "integer",
            "example": 24
          },
          "prev_page_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "to": {
            "type": "integer",
            "nullable": true
          },
          "total": {
            "type": "integer",
            "example": 100
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthenticated."
                }
              }
            }
          }
        }
      },
...
    }
  }
}